Add test for gpg-agent handling.
The only real fragile bit is that gpg-preset-passphrase is installed
in different places on different OSes, and often not in the user path.
For the moment, hard-code a FHS compliant path, and allow the location
to be set by an environment variable.

This is just a copy of the "roundtrip.t" test that already exists,
with agent handling attached to the front. It could probably be made
shorter since some of the tests here do not rely on passphrases.

Also, it might be desirable to re-order the tests.